requirements.txt
asgiref==3.2.10
django==3.1.1
pytz==2020.1
sqlparse==0.3.1
dj-database-url==0.5.0
psycopg2==2.8.5
Faker==4.4.0
factory-boy==3.1.0
python-dateutil==2.8.1
six==1.15.0
text-unidecode==1.3
pycodestyle==2.6.0
importchecker==2.0
cssselect==1.1.0
lxml==4.5.2
pyquery==1.4.1
WebOb==1.8.6
beautifulsoup4==4.9.3
django-webtest==1.9.7
soupsieve==2.0.1
waitress==1.4.4
webtest==2.0.35
python-memcached==1.59
django-cache-memoize==0.1.7
django-nested-inline==0.4.2
certifi==2020.11.8
sentry-sdk==0.19.4
urllib3==1.26.2
bcrypt==3.2.0
cffi==1.14.3
cryptography==3.2.1
fabric==2.5.0
invoke==1.4.1
paramiko==2.7.2
pycparser==2.20
pynacl==1.4.0
GitPython==3.1.11
gitdb==4.0.5
smmap==3.0.4
djangorestframework~=3.12.2
ngram==3.3.2
chardet==4.0.0
extruct==0.12.0
html-text==0.5.2
html5lib==1.1
idna==2.10
isodate==0.6.0
jstyleson==0.0.2
mf2py==1.1.2
pyparsing==2.4.7
pyrdfa3==3.5.3
rdflib==5.0.0
rdflib-jsonld==0.5.0
requests==2.25.1
w3lib==1.22.0
webencodings==0.5.1
inflection==0.5.1
dacite==1.6.0
django-filter==2.4.0
django-crispy-forms==1.11.2